+++ esbuild.js
@@ -35,7 +35,7 @@ async function start() {
     define: {'process.env.NODE_ENV': JSON.stringify("production")},
     bundle: true,
     platform: 'node',
-    target: 'node12.16',
+    target: 'node22',
     mainFields: ['module', 'main'],
     minify: true,
     sourcemap: true,
@@ -47,4 +47,5 @@ async function start() {
 
 start().catch(e => {
   console.error(e)
+  process.exitCode = 1
 })